wechat_analysis_agent_clean_v2/backend/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.29
python-dotenv==0.19.2
requests==2.26.0
volcengine-python-sdk[ark]>=5.0.0